+++ src/__tests__/views/TransactionsPage/TransactionsPage.test.js
@@ -1,4 +1,4 @@
-import { screen, waitFor } from "@testing-library/react";
+import { fireEvent, screen, waitFor } from "@testing-library/react";
 import { renderWithProviders } from "../../__test-utils__/test-utils";
 import { rest } from "msw";
 // eslint-disable-next-line jest/no-mocks-import
@@ -50,6 +50,46 @@ describe("Test TransactionsPage", () => {
     await waitFor(() => {
       const noData = screen.getByText(/There are no records to display/i);
       expect(noData).toBeInTheDocument();
+    });
+  });
+
+  it("Test can manage change page", async () => {
+
+    renderWithProviders(<GenericPage entity={ENTITY_TRANSACTIONS} />);
+    await waitFor(() => {
+      const nextPageButton = screen.getByRole("button", {name: "Next Page"});
+      expect(nextPageButton).toBeInTheDocument();
+      fireEvent.click(nextPageButton);
+    });
+  });
+
+  it("Test can manage change per page", async () => {
+
+    renderWithProviders(<GenericPage entity={ENTITY_TRANSACTIONS} />);
+    await waitFor(() => {
+      const perPageSelect = screen.getByLabelText('Rows per page:');
+      expect(perPageSelect).toBeInTheDocument();
+      fireEvent.change(perPageSelect, {
+        target: { value: "20" },
+      });
+    });
+  });
+
+  it("Test can manage sort", async () => {
+    renderWithProviders(<GenericPage entity={ENTITY_TRANSACTIONS} />);
+    await waitFor(() => {
+      const column = screen.getByText(/amount/i);
+      expect(column).toBeInTheDocument();
+      fireEvent.click(column);
+    });
+  });
+
+  it("Test can manage callback", async () => {
+    renderWithProviders(<GenericPage entity={ENTITY_TRANSACTIONS} forceUndefinedForSort />);
+    await waitFor(() => {
+      const column = screen.getByText(/amount/i);
+      expect(column).toBeInTheDocument();
+      fireEvent.click(column);
     });
   });