+++ src/store/index.js
@@ -1,17 +1,12 @@
 import { configureStore } from "@reduxjs/toolkit";
-import { userSlice } from "../features/userSlice";
-import { loginSlice } from "../features/loginSlice";
-import { balanceSlice } from "../features/balanceSlice";
+import { apiSlice } from "../features/apiSlice";
 import addAccessTokenMiddleware from "../middlewares/addAccessTokenMiddleware";
-import { transactionsSlice } from "../features/transactionsSlice";
+import removeAccessTokenMiddleware from "../middlewares/removeAccessTokenMiddleware";
 
 const setupStore = (preloadedState) =>
   configureStore({
     reducer: {
-      [loginSlice.reducerPath]: loginSlice.reducer,
-      [balanceSlice.reducerPath]: balanceSlice.reducer,
-      [transactionsSlice.reducerPath]: transactionsSlice.reducer,
-      userSlice: userSlice.reducer
+      [apiSlice.reducerPath]: apiSlice.reducer
     },
     preloadedState,
     middleware: (getDefaultMiddleware) =>
@@ -20,9 +15,8 @@ const setupStore = (preloadedState) =>
         serializableCheck: false
       })
         .prepend(addAccessTokenMiddleware.middleware)
-        .concat(loginSlice.middleware)
-        .concat(balanceSlice.middleware)
-        .concat(transactionsSlice.middleware)
+        .prepend(removeAccessTokenMiddleware.middleware)
+        .concat(apiSlice.middleware)
   });
 
 export default setupStore;