+++ CHANGELOG.md
@@ -1,4 +1,10 @@
-# 0.7.0
+# 0.8.0
++ Remove token middleware on logout
++ Remove `usersSlice`
++ Add Footer to HomePage
++ Add pagination
+
+## 0.7.0
 + Remove docker-compose and refactor Docker
 + Upgrade dependencies
 + Remove accessTokenSlice